Prevent denial of service in GXCommon.getCompactArray
If a malfunctioning/rogue device indicated zero columns for a compact array, the code entered an infinite loop, allocating empty ArrayLists until heap memory got exhausted. This could lead to denial-of-service attacks, as the caller has no means for recovery attempt.
This change defensively detects such abnormal situation and handles it by returning a simple list of empty lists.
In addition, formerly overloaded variable len is disambiguated into colCount and rowCount in order to improve clarity of the code.