Fix bulk removal method in GXArray
The original code removed individual list items in ascending index order. However, it deleted only every second element, because it incremented index and shifted elements at the same time. Probability of throwing an exception is high when operating near the end of the list. The code was replaced by the idiomatic approach suggested by JDK Javadoc of List.sublist method.
The interpretation of the boundary indices' behavior - original code implies that both are inclusive - was documented. This interpretation, however, appears incorrect in context of calls from GXDLMSArrayManager, and needs to be double-checked during code review.